first pass of application

+++ assets/grocery_crud/themes/flexigrid/views/list.php
@@ -38,14 +38,15 @@
 			<?php if(!$unset_delete || !$unset_edit || !empty($actions)){?>
 			<td align="left" width='20%'>
 				<div class='tools'>				
-					<?php if(!$unset_delete){?>
+					<?php if(!$unset_delete && $row->delete_url){?>
                     	<a href='<?php echo $row->delete_url?>' title='<?php echo $this->l('list_delete')?> <?php echo $subject?>' class="delete-row" >
                     			<span class='delete-icon'></span>
                     	</a>
                     <?php }?>
-                    <?php if(!$unset_edit){?>
+                    <?php if(!$unset_edit && $row->edit_url){?>
 						<a href='<?php echo $row->edit_url?>' title='<?php echo $this->l('list_edit')?> <?php echo $subject?>'><span class='edit-icon'></span></a>
 					<?php }?>
+						<a href='<?php echo $row->view_url?>' title='<?php echo $this->l('list_view')?> <?php echo $subject?>'><span class='view-icon'></span></a>
 					<?php 
 					if(!empty($row->action_urls)){
 						foreach($row->action_urls as $action_unique_id => $action_url){
